+++ polaris_gem_experiments/prog-analysis/lane_centering_stanley/main.c
@@ -1,6 +1,6 @@
 #include <assert.h>
 
-#include "stanley_lane_centering.h"
+#include "lane_centering_stanley.h"
 
 #define HALF_LANE_W ((ApproxReal) 0.38)
